requirements.txt
appdirs==1.4.4
asgiref==3.2.7
Babel==2.8.0
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
cmake==3.15.3
cryptography==3.1.1
cycler==0.10.0
decorator==4.4.1
defusedxml==0.7.0rc1
distlib==0.3.0
Django==3.0.6
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
django-sslserver==0.22
djangorestframework==3.11.0
dlib==19.18.0
et-xmlfile==1.0.1
facebook-sdk==3.1.0
filelock==3.0.12
googletrans==2.4.0
idna==2.8
imageio==2.6.1
imutils==0.5.3
jdcal==1.4.1
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.3
mysql-connector-python==8.0.18
networkx==2.4
numpy==1.17.4
oauthlib==3.1.0
object-detection==0.1
opencv-contrib-python==4.2.0.34
opencv-python==3.4.8.29
openpyxl==3.0.3
Pillow==6.2.1
protobuf==3.11.1
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python3-openid==3.2.0
requests==2.22.0
requests-oauthlib==1.3.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.7
virtualenv==20.0.21
